requirements.txt
asttokens==2.2.1
backcall==0.2.0
catkin==0.7.29
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
fonttools==4.41.1
gym==0.26.2
gym-notices==0.0.8
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard-logger==0.1.0
threadpoolctl==3.2.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
